Adding an action block to the workflow file
Workflows piece together jobs, and jobs piece together steps.
With this commit we are now creating a job that runs an action.
Actions can be used from within the same repository, from any
other public repository, or from a published Docker container
image.

We are using an action that will be defined in this repository.